+++ packages/scripts/utils/index.js
@@ -13,6 +13,7 @@ const {
 } = require( './cli' );
 const {
 	getJestOverrideConfigFile,
+	getVitestOverrideConfigFile,
 	getPhpFilePaths,
 	getProjectSourcePath,
 	getWebpackArgs,
@@ -20,6 +21,7 @@ const {
 	hasBabelConfig,
 	hasCssnanoConfig,
 	hasJestConfig,
+	hasVitestConfig,
 	hasPostCSSConfig,
 	hasPrettierConfig,
 } = require( './config' );
@@ -44,6 +46,7 @@ module.exports = {
 	getArgsFromCLI,
 	getFileArgsFromCLI,
 	getJestOverrideConfigFile,
+	getVitestOverrideConfigFile,
 	getNodeArgsFromCLI,
 	getPackageProp,
 	getPhpFilePaths,
@@ -57,6 +60,7 @@ module.exports = {
 	hasCssnanoConfig,
 	hasFileArgInCLI,
 	hasJestConfig,
+	hasVitestConfig,
 	hasPackageProp,
 	hasPostCSSConfig,
 	hasPrettierConfig,